refactor(lifecycle): move PR git artifacts into application
Activation still mounts the scheduler in the daemon; Git identity, leases, and teardown proofs now live with pr_tracking.
