+++ crates/tracedecay-application/Cargo.toml
@@ -45,6 +45,7 @@ test-transport = [
 
 [dependencies]
 cap-std = "4.0.2"
+fs2 = "0.4"
 getrandom = "0.2"
 gix = { version = "=0.86.0", default-features = false, features = ["revision", "blob-diff", "parallel", "sha1", "sha256", "status"] }
 glob = "0.3"
