+++ crates/tracedecay-application/src/pr_tracking.rs
@@ -1,5 +1,18 @@
 //! Git-backed pull-request discovery and durable managed-PR state.
 
+mod lifecycle;
+
+pub use lifecycle::{
+    ManualBranchActivation, ManualBranchActivationError, ManualBranchArtifactOwnershipV1,
+    ManualBranchArtifactsV1, ManualBranchLifecycleLeaseV1, PrCleanupArtifact, PrCleanupError,
+    PrCleanupReceipt, ReconcileReport, checkout_linked_worktree, cleanup_owned_worktree,
+    cleanup_owned_worktree_off_runtime, cleanup_pr_worktree, cleanup_pr_worktree_off_runtime,
+    manual_branch_artifact_ownership, manual_branch_artifact_ownership_off_runtime,
+    manual_branch_artifacts_match, manual_branch_artifacts_match_off_runtime,
+    manual_branch_source_owns_artifacts, prepare_manual_branch_worktree, prepare_pr_worktree,
+    ref_points_to, remove_worktree, resolve_branch_head, try_acquire_manual_branch_lifecycle,
+};
+
 use std::collections::{BTreeMap, HashMap};
 use std::path::{Path, PathBuf};
 use std::process::ExitStatus;
