+++ crates/tracedecay-application/tests/pr_tracking.rs
@@ -50,6 +50,55 @@ fn managed_pr_state_round_trips_through_application_owner() {
 }
 
 #[test]
+fn manual_branch_artifacts_hash_the_branch_name_not_the_raw_path() {
+    let store = tempfile::tempdir().expect("store root");
+    let slashed = tracedecay_application::pr_tracking::ManualBranchArtifactsV1::for_branch(
+        store.path(),
+        "feature/a",
+    );
+    let underscored = tracedecay_application::pr_tracking::ManualBranchArtifactsV1::for_branch(
+        store.path(),
+        "feature_a",
+    );
+
+    assert_ne!(slashed.worktree, underscored.worktree);
+    assert_ne!(slashed.tracking_ref, underscored.tracking_ref);
+    assert!(
+        slashed
+            .worktree
+            .starts_with(store.path().join("branch-worktrees"))
+    );
+}
+
+#[test]
+fn manual_branch_lifecycle_lease_is_exclusive_per_branch() {
+    let store = tempfile::tempdir().expect("store root");
+    let first = tracedecay_application::pr_tracking::try_acquire_manual_branch_lifecycle(
+        store.path(),
+        "feature/exclusive",
+    )
+    .expect("first lease");
+    let contended = tracedecay_application::pr_tracking::try_acquire_manual_branch_lifecycle(
+        store.path(),
+        "feature/exclusive",
+    );
+
+    assert!(first.matches_branch("feature/exclusive"));
+    assert!(matches!(
+        contended,
+        Err(
+            tracedecay_application::pr_tracking::ManualBranchActivationError::LifecycleContended { .. }
+        )
+    ));
+    drop(first);
+    tracedecay_application::pr_tracking::try_acquire_manual_branch_lifecycle(
+        store.path(),
+        "feature/exclusive",
+    )
+    .expect("lease after drop");
+}
+
+#[test]
 fn malformed_managed_pr_state_is_a_typed_json_error() {
     let store = tempfile::tempdir().expect("store root");
     std::fs::write(store.path().join("pr-autotrack.json"), "{not json")
